refactor into smaller components; helper render functions
