+++ services/QuillLMS/client/app/bundles/Evidence/components/studentView/steps.tsx
@@ -1,0 +1,42 @@
+import * as React from 'react';
+
+import { renderReadPassageStep, renderPromptSteps } from '../../helpers/containerRenderHelpers';
+
+const Steps = ({
+  activeStep,
+  activateStep,
+  activities,
+  handleDoneReadingClick,
+  completeStep,
+  submitResponse,
+  closeReadTheDirectionsModal,
+  session,
+  completedSteps,
+  doneHighlighting,
+  resetTimers,
+  showReadTheDirectionsModal,
+  stepsHash
+}) => {
+  return(
+    <div className="steps-outer-container" onScroll={resetTimers}>
+      <div className="steps-inner-container" onScroll={resetTimers}>
+        {renderReadPassageStep(activeStep, activities, handleDoneReadingClick)}
+        {renderPromptSteps({
+          activateStep,
+          completeStep,
+          submitResponse,
+          closeReadTheDirectionsModal,
+          activities,
+          session,
+          activeStep,
+          completedSteps,
+          doneHighlighting,
+          showReadTheDirectionsModal,
+          stepsHash
+        })}
+      </div>
+    </div>
+  )
+}
+
+export default Steps;
